requirements.txt
aiohttp==2.2.5
async-timeout==2.0.0
cffi==1.11.2
chardet==3.0.4
discord.py==1.0.0a0
multidict==3.3.2
pycparser==2.18
PyNaCl==1.1.2
six==1.11.0
websockets==3.4
yarl==0.16.0
